Extract processing of `by` option into a dedicated class
